refactor(github): improve PR detection and pagination
- Use `octokit.paginate` for `listOpenPullRequests()` and `getPullRequestFiles()`
- Add `getCurrentUser()` and `getForkOwner()` methods
- Add `createdBy` and `lastCommitAuthor` to PR status validation
- Remove branch-based PR detection in favor of file-based only
- Move markdown filtering to repository level
- Simplify PR title generation in locale files
- Make `prNumber` required in `CommentBuilderService`
- Refactor runner managers into grouped object structure
- Add translator manager modules for chunks and validation
