+++ src/locales/ru.locale.ts
@@ -78,8 +78,7 @@ export const ruLocale: LocaleDefinition = {
 - Preserve English technical terms that are commonly used untranslated in Russian developer communities (e.g., "render", "props", "state", "hook")`,
 	},
 	pullRequest: {
-		title: (file: TranslationFile) =>
-			`Перевод ${file.title ? `**${file.title}**` : `\`${file.filename}\``} на русский язык`,
+		title: (file: TranslationFile) => `Перевод \`${file.filename}\` на русский язык`,
 		body: createPRBodyBuilder(ruPRBodyStrings),
 	},
 };
