+++ src/services/comment-builder/comment-builder.service.ts
@@ -9,7 +9,7 @@ import { TranslationFile } from "../translator/translator.service";
 
 export interface FileEntry {
 	file: TranslationFile;
-	prNumber: number | undefined;
+	prNumber: number;
 }
 
 export interface HierarchicalStructure {
@@ -57,6 +57,8 @@ export class CommentBuilderService {
 	public buildComment(results: ProcessedFileResult[], filesToTranslate: TranslationFile[]) {
 		const concattedData = results
 			.map((result) => {
+				if (!result.pullRequest) return null;
+
 				const translationFile = filesToTranslate.find((file) => file.filename === result.filename);
 
 				if (!translationFile) return null;
@@ -66,7 +68,7 @@ export class CommentBuilderService {
 				return {
 					file: translationFile,
 					pathParts: this.simplifyPathParts(pathParts),
-					prNumber: result.pullRequest?.number,
+					prNumber: result.pullRequest.number,
 				} satisfies FileWithHierarchy;
 			})
 			.filter(Boolean);
@@ -224,7 +226,7 @@ export class CommentBuilderService {
 
 		const dirs = Object.keys(structure)
 			.filter((key) => key !== "files")
-			.sort();
+			.toSorted();
 
 		for (const dir of dirs) {
 			const currentLevel = structure[dir];
@@ -232,9 +234,9 @@ export class CommentBuilderService {
 				continue;
 			}
 
-			const sortedEntries = currentLevel.files.toSorted((left, right) =>
-				left.file.filename.localeCompare(right.file.filename),
-			);
+			const sortedEntries = currentLevel.files.toSorted((left, right) => {
+				return left.file.filename.localeCompare(right.file.filename);
+			});
 
 			if (
 				sortedEntries.length === 1 &&
@@ -248,8 +250,6 @@ export class CommentBuilderService {
 			lines.push(`${indent}- ${dir}`);
 
 			for (const entry of sortedEntries) {
-				if (!entry.prNumber) continue;
-
 				lines.push(`${indent}  - #${entry.prNumber}`);
 			}
