+++ src/services/translator/managers/index.ts
@@ -1,0 +1,2 @@
+export * from "./translation-validator.manager";
+export * from "./chunks.manager";
