+++ tests/services/github/github.service.spec.ts
@@ -88,7 +88,7 @@ describe("GitHubService", () => {
 
 		describe("getRepositoryTree", () => {
 			test("should return fork tree when target is fork", async () => {
-				const tree = await githubService.getRepositoryTree("fork", "main", false);
+				const tree = await githubService.getRepositoryTree("fork", "main");
 
 				expect(octokitMock.git.getTree).toHaveBeenCalledWith({
 					...testRepositories.fork,
@@ -99,7 +99,7 @@ describe("GitHubService", () => {
 			});
 
 			test("should return upstream tree when target is upstream", async () => {
-				const tree = await githubService.getRepositoryTree("upstream", "main", false);
+				const tree = await githubService.getRepositoryTree("upstream", "main");
 
 				expect(octokitMock.git.getTree).toHaveBeenCalledWith({
 					...testRepositories.upstream,
@@ -110,7 +110,7 @@ describe("GitHubService", () => {
 			});
 
 			test("should default to fork when no target is specified", async () => {
-				const tree = await githubService.getRepositoryTree(undefined, "main", false);
+				const tree = await githubService.getRepositoryTree(undefined, "main");
 
 				expect(octokitMock.git.getTree).toHaveBeenCalledWith({
 					...testRepositories.fork,
@@ -123,7 +123,7 @@ describe("GitHubService", () => {
 			test("should fetch default branch when base branch is not specified", async () => {
 				octokitMock.repos.get.mockResolvedValueOnce({ data: { default_branch: "develop" } });
 
-				await githubService.getRepositoryTree("fork", undefined, false);
+				await githubService.getRepositoryTree("fork", undefined);
 
 				expect(octokitMock.repos.get).toHaveBeenCalledWith(testRepositories.fork);
 				expect(octokitMock.git.getTree).toHaveBeenCalledWith({
@@ -149,21 +149,6 @@ describe("GitHubService", () => {
 
 				expect(tree).toHaveLength(1);
 				expect(tree[0]?.path).toBe("src/test/file.md");
-			});
-
-			test("should not filter tree when filterIgnored is false", async () => {
-				octokitMock.git.getTree.mockResolvedValueOnce({
-					data: {
-						tree: [
-							{ path: "src/test/file.md", type: "blob", sha: "abc123", url: "", mode: "100644" },
-							{ path: "README.md", type: "blob", sha: "def456", url: "", mode: "100644" },
-						],
-					},
-				});
-
-				const tree = await githubService.getRepositoryTree("fork", "main", false);
-
-				expect(tree).toHaveLength(2);
 			});
 
 			test("should throw RequestError when API call fails", () => {
@@ -671,19 +656,18 @@ describe("GitHubService", () => {
 
 		describe("listOpenPullRequests", () => {
 			test("should return list of open pull requests when called", async () => {
-				octokitMock.pulls.list.mockResolvedValueOnce({
-					data: [
-						{ number: 1, title: "PR 1" },
-						{ number: 2, title: "PR 2" },
-					],
-				} as RestEndpointMethodTypes["pulls"]["list"]["response"]);
+				octokitMock.paginate.mockResolvedValueOnce([
+					{ number: 1, title: "PR 1" },
+					{ number: 2, title: "PR 2" },
+				] as RestEndpointMethodTypes["pulls"]["list"]["response"]["data"]);
 
 				const prs = await githubService.listOpenPullRequests();
 
 				expect(prs).toHaveLength(2);
-				expect(octokitMock.pulls.list).toHaveBeenCalledWith({
+				expect(octokitMock.paginate).toHaveBeenCalledWith("GET /repos/{owner}/{repo}/pulls", {
 					...testRepositories.upstream,
 					state: "open",
+					per_page: 100,
 				});
 			});
 
@@ -697,7 +681,7 @@ describe("GitHubService", () => {
 
 			test("should throw RequestError when API call fails", () => {
 				const apiError = createOctokitRequestErrorFixture({ message: "API Error" });
-				octokitMock.pulls.list.mockRejectedValueOnce(apiError);
+				octokitMock.paginate.mockRejectedValueOnce(apiError);
 
 				expect(githubService.listOpenPullRequests()).rejects.toThrow(apiError);
 			});
@@ -705,21 +689,22 @@ describe("GitHubService", () => {
 
 		describe("getPullRequestFiles", () => {
 			test("should return list of file paths when PR has files", async () => {
-				octokitMock.pulls.listFiles.mockResolvedValueOnce({
-					data: [{ filename: "src/file1.md" }, { filename: "src/file2.md" }],
-				} as RestEndpointMethodTypes["pulls"]["listFiles"]["response"]);
+				octokitMock.paginate.mockResolvedValueOnce([
+					{ filename: "src/file1.md" },
+					{ filename: "src/file2.md" },
+				]);
 
 				const files = await githubService.getPullRequestFiles(123);
 
 				expect(files).toEqual(["src/file1.md", "src/file2.md"]);
-				expect(octokitMock.pulls.listFiles).toHaveBeenCalledWith({
-					...testRepositories.upstream,
-					pull_number: 123,
-				});
+				expect(octokitMock.paginate).toHaveBeenCalledWith(
+					"GET /repos/{owner}/{repo}/pulls/{pull_number}/files",
+					{ ...testRepositories.upstream, pull_number: 123, per_page: 100 },
+				);
 			});
 
 			test("should return empty array when PR has no files", async () => {
-				octokitMock.pulls.listFiles.mockResolvedValueOnce({ data: [] });
+				octokitMock.paginate.mockResolvedValueOnce([]);
 
 				const files = await githubService.getPullRequestFiles(123);
 
@@ -728,7 +713,7 @@ describe("GitHubService", () => {
 
 			test("should throw RequestError when API call fails", () => {
 				const apiError = createOctokitRequestErrorFixture({ message: "API Error" });
-				octokitMock.pulls.listFiles.mockRejectedValueOnce(apiError);
+				octokitMock.paginate.mockRejectedValueOnce(apiError);
 
 				expect(githubService.getPullRequestFiles(123)).rejects.toThrow(apiError);
 			});
@@ -801,6 +786,7 @@ describe("GitHubService", () => {
 							number: 1,
 							mergeable: true,
 							mergeable_state: "clean",
+							user: { login: "test-user" },
 						},
 					}),
 				);
@@ -810,11 +796,17 @@ describe("GitHubService", () => {
 				expect(status.hasConflicts).toBe(false);
 				expect(status.needsUpdate).toBe(false);
 				expect(status.mergeableState).toBe("clean");
+				expect(status.createdBy).toBe("test-user");
 			});
 
 			test("should return dirty status when PR has conflicts", async () => {
 				octokitMock.pulls.get.mockResolvedValueOnce({
-					data: { number: 1, mergeable: false, mergeable_state: "dirty" },
+					data: {
+						number: 1,
+						mergeable: false,
+						mergeable_state: "dirty",
+						user: { login: "other-user" },
+					},
 				});
 
 				const status = await githubService.checkPullRequestStatus(1);
@@ -822,11 +814,17 @@ describe("GitHubService", () => {
 				expect(status.hasConflicts).toBe(true);
 				expect(status.needsUpdate).toBe(true);
 				expect(status.mergeableState).toBe("dirty");
+				expect(status.createdBy).toBe("other-user");
 			});
 
 			test("should return behind status without conflicts", async () => {
 				octokitMock.pulls.get.mockResolvedValueOnce({
-					data: { number: 1, mergeable: true, mergeable_state: "behind" },
+					data: {
+						number: 1,
+						mergeable: true,
+						mergeable_state: "behind",
+						user: { login: "test-fork-owner" },
+					},
 				});
 
 				const status = await githubService.checkPullRequestStatus(1);
@@ -834,6 +832,22 @@ describe("GitHubService", () => {
 				expect(status.hasConflicts).toBe(false);
 				expect(status.needsUpdate).toBe(false);
 				expect(status.mergeableState).toBe("behind");
+				expect(status.createdBy).toBe("test-fork-owner");
+			});
+
+			test("should return creator login when user exists", async () => {
+				octokitMock.pulls.get.mockResolvedValueOnce({
+					data: {
+						number: 1,
+						mergeable: true,
+						mergeable_state: "clean",
+						user: { login: "test-creator" },
+					},
+				});
+
+				const status = await githubService.checkPullRequestStatus(1);
+
+				expect(status.createdBy).toBe("test-creator");
 			});
 
 			test("should throw RequestError when API call fails", () => {
@@ -841,6 +855,27 @@ describe("GitHubService", () => {
 				octokitMock.pulls.get.mockImplementation(() => Promise.reject(apiError));
 
 				expect(githubService.checkPullRequestStatus(1)).rejects.toThrow(apiError);
+			});
+		});
+
+		describe("getForkOwner", () => {
+			test("should return fork owner from repositories config", () => {
+				const owner = githubService.getForkOwner();
+
+				expect(owner).toBe("test-fork-owner");
+			});
+
+			test("should return custom fork owner when overridden", () => {
+				const customService = createTestGitHubService({
+					repositories: {
+						fork: { owner: "custom-owner", repo: "custom-repo" },
+						upstream: testRepositories.upstream,
+					},
+				});
+
+				const owner = customService.getForkOwner();
+
+				expect(owner).toBe("custom-owner");
 			});
 		});
