+++ tests/services/translator.service.spec.ts
@@ -195,7 +195,8 @@ describe("TranslatorService", () => {
 
 			const file = createTranslationFileFixture({ content: "Conteúdo em português." });
 
-			const analysis = await translatorService.getLanguageAnalysis(file);
+			const analysis =
+				await translatorService.managers.translationValidator.getLanguageAnalysis(file);
 
 			expect(analysis).toBe(expectedAnalysis);
 			expect(analyzeSpy).toHaveBeenCalledWith(file.filename, file.content);
@@ -204,7 +205,9 @@ describe("TranslatorService", () => {
 		test("throws ApplicationError when file content is empty", () => {
 			const file = createTranslationFileFixture({ content: "" });
 
-			expect(translatorService.getLanguageAnalysis(file)).rejects.toThrow("File content is empty");
+			expect(
+				translatorService.managers.translationValidator.getLanguageAnalysis(file),
+			).rejects.toThrow("File content is empty");
 		});
 	});
 
@@ -304,7 +307,8 @@ describe("TranslatorService", () => {
 				content: "This is English content that needs translation.",
 			});
 
-			const isTranslated = await translatorService.isContentTranslated(file);
+			const isTranslated =
+				await translatorService.managers.translationValidator.isContentTranslated(file);
 
 			expect(isTranslated).toBe(false);
 		});
@@ -318,7 +322,8 @@ describe("TranslatorService", () => {
 				content: "Este é um conteúdo em português que já foi traduzido.",
 			});
 
-			const isTranslated = await translatorService.isContentTranslated(file);
+			const isTranslated =
+				await translatorService.managers.translationValidator.isContentTranslated(file);
 
 			expect(isTranslated).toBe(true);
 		});
@@ -328,7 +333,8 @@ describe("TranslatorService", () => {
 				content: "This has some English and também algum português.",
 			});
 
-			const isTranslated = await translatorService.isContentTranslated(file);
+			const isTranslated =
+				await translatorService.managers.translationValidator.isContentTranslated(file);
 
 			expect(typeof isTranslated).toBe("boolean");
 		});
@@ -340,7 +346,8 @@ describe("TranslatorService", () => {
 
 			const file = createTranslationFileFixture({ content: "Some content." });
 
-			const isTranslated = await translatorService.isContentTranslated(file);
+			const isTranslated =
+				await translatorService.managers.translationValidator.isContentTranslated(file);
 
 			expect(isTranslated).toBe(false);
 		});
